image/requirements.txt
alpaca-py
boto3
requests
decimal